test: extract shared resolveBundledAssetUri helper
Pull asset resolution + file:// stripping out of the duplicated copies in
mobile/consumer.ts and mobile/executors/model-asset-executor.ts into a
single mobile/asset-uri.ts helper. Both sites now delegate, so future
changes to expo-asset handling live in one place.

Tighter idioms in the helper itself: regex strip instead of substring(7),
?? instead of ||, no mutable let.
